requirements.txt
qrcode>=7.3.1
pillow>=9.0.0
pyzbar>=0.1.9
cryptography>=38.0.0
